requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
matplotlib>=3.5.0
python-dotenv>=1.0.0
requests>=2.28.0
openai>=0.27.0
transformers>=4.30.0
scikit-learn>=1.0.0
nltk>=3.7
beautifulsoup4>=4.10.0
SQLAlchemy>=2.0.0 
seaborn>=0.13.0
nltk>=3.8.1
spacy>=3.8.0
yake>=0.4.8
wordcloud>=1.9.0
readability-lxml>=0.8.4
